Split config file reading and generation to separate files
